Update 1.5.2 - Code quality stuffs

+++ src/main/java/com/kuborros/FurBotNeo/commands/MusicCommands/MusicTimeCommand.java
@@ -60,9 +60,9 @@ public class MusicTimeCommand extends MusicCommand {
             } else {
                 val = val.trim();
             }
-            int seconds;
+            long seconds;
             try {
-                seconds = (min ? 60 : 1) * Integer.parseInt(val);
+                seconds = (min ? 60 : 1) * Long.parseLong(val);
                 long milis = (seconds * 1000);
                 long duration = getPlayer(guild).getPlayingTrack().getDuration();
                 if (duration <= milis) {
